restructuring files, made UserInterface component
