+++ client/components/index.js
@@ -3,6 +3,7 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Main} from './main'
+
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
+export {UserInterface} from './UserInterface.jsx'
