+++ client/index.js
@@ -2,14 +2,14 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import {Provider} from 'react-redux'
 import store from './store'
-import Routes from './routes'
+import UserInterface from './components'
 
 // establishes socket connection
 import './socket'
 
 ReactDOM.render(
   <Provider store={store}>
-    <Routes />
+    <UserInterface />
   </Provider>,
   document.getElementById('app')
 )
